The usual stuff

+++ include/charon/EnergyStencils/EnergyNonLocal.hxx
@@ -145,11 +145,11 @@ T EnergyNonLocal<T>::getEnergy( int, int xI, int yI, int zI, int )
       weight = spatial_weight * color_weight * occlusion_weight;
       weight_sum += weight;
 
-      pixelEnergy += weight * _penaltyFunction->getPenalty( fabs(du) );
-      pixelEnergy += weight * _penaltyFunction->getPenalty( fabs(dv) );
+      pixelEnergy += weight * _penaltyFunction->getPenalty( fabs(double(du)) );
+      pixelEnergy += weight * _penaltyFunction->getPenalty( fabs(double(dv)) );
     } else {
-      pixelEnergy += _penaltyFunction->getPenalty( fabs(du) );
-      pixelEnergy += _penaltyFunction->getPenalty( fabs(dv) );
+      pixelEnergy += _penaltyFunction->getPenalty( fabs(double(du)) );
+      pixelEnergy += _penaltyFunction->getPenalty( fabs(double(dv)) );
       weight_sum += 1.0;
     }
   }
@@ -222,11 +222,11 @@ std::vector<T> EnergyNonLocal<T>::getEnergyGradient(
       weight = spatial_weight * color_weight * occlusion_weight;
       weight_sum += weight;
 
-      pixelGradientU += weight * _penaltyFunction->getPenaltyGradient( fabs(du) ) * signum(du);
-      pixelGradientV += weight * _penaltyFunction->getPenaltyGradient( fabs(dv) ) * signum(dv);
+      pixelGradientU += weight * _penaltyFunction->getPenaltyGradient( fabs(double(du)) ) * signum(du);
+      pixelGradientV += weight * _penaltyFunction->getPenaltyGradient( fabs(double(dv)) ) * signum(dv);
     } else {
-      pixelGradientU += _penaltyFunction->getPenaltyGradient( fabs(du) ) * signum(du);
-      pixelGradientV += _penaltyFunction->getPenaltyGradient( fabs(dv) ) * signum(dv);
+      pixelGradientU += _penaltyFunction->getPenaltyGradient( fabs(double(du)) ) * signum(double(du));
+      pixelGradientV += _penaltyFunction->getPenaltyGradient( fabs(double(dv)) ) * signum(double(dv));
       weight_sum += 1.0;
     }
   }
